+++ swak/pt/embed/gated.py
@@ -34,6 +34,9 @@ class GatedEmbedder(Module):
     ) -> None:
         super().__init__()
         self.mod_dim = mod_dim
+        # Although few, some activation functions have learnable parameters
+        if hasattr(gate, 'reset_parameters'):
+            gate.reset_parameters()
         self.gate = gate
         self.inp_dim = inp_dim
         self.kwargs = kwargs
@@ -64,6 +67,9 @@ class GatedEmbedder(Module):
     def reset_parameters(self) -> None:
         """Re-initialize all internal parameters."""
         self.embed.reset_parameters()
+        # Although few, some activation functions have learnable parameters
+        if hasattr(self.gate, 'reset_parameters'):
+            self.gate.reset_parameters()
 
     def new(
             self,