+++ swak/pt/mix/weighted/gated_residual.py
@@ -71,6 +71,11 @@ class GatedResidualSumMixer(Module):
         super().__init__()
         self.mod_dim = mod_dim
         self.n_features = n_features
+        # Although few, some activation functions have learnable parameters
+        if hasattr(activate, 'reset_parameters'):
+            activate.reset_parameters()
+        if hasattr(gate, 'reset_parameters'):
+            gate.reset_parameters()
         self.activate = activate
         self.gate = gate
         self.drop = drop
@@ -128,6 +133,11 @@ class GatedResidualSumMixer(Module):
         """Re-initialize all internal parameters."""
         self.project.reset_parameters()
         self.widen.reset_parameters()
+        # Although few, some activation functions have learnable parameters
+        if hasattr(self.activate, 'reset_parameters'):
+            self.activate.reset_parameters()
+        if hasattr(self.gate, 'reset_parameters'):
+            self.gate.reset_parameters()
 
     def new(
             self,