+++ tests/pt/blocks/test_activated_gated_block.py
@@ -2,14 +2,14 @@ import unittest
 from unittest.mock import patch, Mock
 import torch as pt
 import torch.nn as ptn
-from swak.pt.blocks import GatedResidualBlock
+from swak.pt.blocks import ActivatedGatedBlock
 from swak.pt.misc import identity
 
 
 class TestDefaultAttributes(unittest.TestCase):
 
     def setUp(self):
-        self.block = GatedResidualBlock(4)
+        self.block = ActivatedGatedBlock(4)
 
     def test_has_mod_dim(self):
         self.assertTrue(hasattr(self.block, 'mod_dim'))
@@ -36,20 +36,18 @@ class TestDefaultAttributes(unittest.TestCase):
     def test_drop(self):
         self.assertIsInstance(self.block.drop, ptn.Dropout)
 
+    def test_has_hidden_factor(self):
+        self.assertTrue(hasattr(self.block, 'hidden_factor'))
+
+    def test_hidden_factor(self):
+        self.assertIsInstance(self.block.hidden_factor, int)
+        self.assertEqual(4, self.block.hidden_factor)
+
     def test_has_kwargs(self):
         self.assertTrue(hasattr(self.block, 'kwargs'))
 
     def test_kwargs(self):
         self.assertDictEqual({}, self.block.kwargs)
-
-    def test_has_project(self):
-        self.assertTrue(hasattr(self.block, 'project'))
-
-    def test_project(self):
-        self.assertIsInstance(self.block.project, ptn.Linear)
-        self.assertEqual(4, self.block.project.in_features)
-        self.assertEqual(4, self.block.project.out_features)
-        self.assertIsInstance(self.block.project.bias, pt.Tensor)
 
     def test_has_widen(self):
         self.assertTrue(hasattr(self.block, 'widen'))
@@ -57,8 +55,17 @@ class TestDefaultAttributes(unittest.TestCase):
     def test_widen(self):
         self.assertIsInstance(self.block.widen, ptn.Linear)
         self.assertEqual(4, self.block.widen.in_features)
-        self.assertEqual(8, self.block.widen.out_features)
+        self.assertEqual(16, self.block.widen.out_features)
         self.assertIsInstance(self.block.widen.bias, pt.Tensor)
+
+    def test_has_shrink(self):
+        self.assertTrue(hasattr(self.block, 'shrink'))
+
+    def test_shrink(self):
+        self.assertIsInstance(self.block.shrink, ptn.Linear)
+        self.assertEqual(16, self.block.shrink.in_features)
+        self.assertEqual(8, self.block.shrink.out_features)
+        self.assertIsInstance(self.block.shrink.bias, pt.Tensor)
 
     def test_has_reset_parameters(self):
         self.assertTrue(hasattr(self.block, 'reset_parameters'))
@@ -71,7 +78,7 @@ class TestDefaultAttributes(unittest.TestCase):
         activate = ptn.PReLU()
         gate = ptn.PReLU()
         with patch('torch.nn.PReLU.reset_parameters') as mock:
-            _ = GatedResidualBlock(4, activate, gate)
+            _ = ActivatedGatedBlock(4, activate, gate)
             self.assertEqual(2, mock.call_count)
             self.assertEqual(2, linear.call_count)
 
@@ -80,8 +87,8 @@ class TestDefaultAttributes(unittest.TestCase):
         self.block.reset_parameters()
         self.assertEqual(2, mock.call_count)
 
-    def test_reset_parameters_called_on_activation(self):
-        block = GatedResidualBlock(4, ptn.PReLU(), ptn.PReLU())
+    def test_reset_parameters_called_on_activations(self):
+        block = ActivatedGatedBlock(4, ptn.PReLU(), ptn.PReLU())
         with patch('torch.nn.PReLU.reset_parameters') as activation:
             block.reset_parameters()
             self.assertEqual(2, activation.call_count)
@@ -94,22 +101,24 @@ class TestDefaultAttributes(unittest.TestCase):
 
     def test_call_new(self):
         new = self.block.new()
-        self.assertIsInstance(new, GatedResidualBlock)
+        self.assertIsInstance(new, ActivatedGatedBlock)
         self.assertEqual(self.block.mod_dim, new.mod_dim)
         self.assertIs(self.block.activate, new.activate)
         self.assertIs(self.block.gate, new.gate)
         self.assertIs(self.block.drop, new.drop)
+        self.assertEqual(self.block.hidden_factor, new.hidden_factor)
         self.assertDictEqual(self.block.kwargs, new.kwargs)
 
 
 class TestAttributes(unittest.TestCase):
 
     def setUp(self):
-        self.block = GatedResidualBlock(
+        self.block = ActivatedGatedBlock(
             4,
             ptn.ReLU(),
             ptn.GELU(),
-            ptn.AlphaDropout(),
+            ptn.AlphaDropout(0.1),
+            2,
             bias=False
         )
 
@@ -122,19 +131,26 @@ class TestAttributes(unittest.TestCase):
     def test_drop(self):
         self.assertIsInstance(self.block.drop, ptn.AlphaDropout)
 
+    def test_hidden_factor(self):
+        self.assertIsInstance(self.block.hidden_factor, int)
+        self.assertEqual(2, self.block.hidden_factor)
+
     def test_kwargs(self):
         self.assertDictEqual({'bias': False}, self.block.kwargs)
 
     def test_widen(self):
         self.assertIsNone(self.block.widen.bias)
 
+    def test_shrink(self):
+        self.assertIsNone(self.block.shrink.bias)
+
 
 class TestUsage(unittest.TestCase):
 
     def setUp(self):
-        self.block = GatedResidualBlock(4, identity, identity, bias=False)
-        self.block.project.weight.data = pt.ones(4, 4)
-        self.block.widen.weight.data = pt.ones(8, 4)
+        self.block = ActivatedGatedBlock(4, identity, identity, bias=False)
+        self.block.widen.weight.data = pt.ones(16, 4)
+        self.block.shrink.weight.data = pt.ones(8, 16)
 
     def test_callable(self):
         self.assertTrue(callable(self.block))
@@ -142,77 +158,59 @@ class TestUsage(unittest.TestCase):
     def test_1d(self):
         inp = pt.ones(4)
         actual = self.block(inp)
-        expected = pt.ones(4) * 130
+        expected = pt.ones(4) * 4096
         pt.testing.assert_close(actual, expected)
 
     def test_2d(self):
         inp = pt.ones(3, 4)
         actual = self.block(inp)
-        expected = pt.ones(3, 4) * 130
+        expected = pt.ones(3, 4) * 4096
         pt.testing.assert_close(actual, expected)
 
     def test_3d(self):
         inp = pt.ones(2, 3, 4)
         actual = self.block(inp)
-        expected = pt.ones(2, 3, 4) * 130
+        expected = pt.ones(2, 3, 4) * 4096
         pt.testing.assert_close(actual, expected)
 
     def test_4d(self):
         inp = pt.ones(5, 2, 3, 4)
         actual = self.block(inp)
-        expected = pt.ones(5, 2, 3, 4) * 130
+        expected = pt.ones(5, 2, 3, 4) * 4096
         pt.testing.assert_close(actual, expected)
 
     def test_empty_dims(self):
         inp = pt.ones(3, 0, 4)
         actual = self.block(inp)
-        expected = pt.ones(3, 0, 4) * 130
+        expected = pt.ones(3, 0, 4) * 4096
         pt.testing.assert_close(actual, expected)
 
-    def test_project_called(self):
-        with patch.object(
-                self.block.project,
-                'forward',
-                return_value=pt.ones(4)
-        ) as mock:
-            expected = pt.ones(4)
-            _ = self.block(expected)
-            mock.assert_called_once_with(expected)
+    def test_widen_called(self):
+        mock = Mock(return_value=pt.ones(16))
+        self.block.widen.forward = mock
+        expected = pt.ones(4)
+        _ = self.block(expected)
+        mock.assert_called_once_with(expected)
 
     def test_activate_called(self):
-        mock = Mock(return_value=pt.ones(4))
+        mock = Mock(return_value=pt.ones(16))
         self.block.activate = mock
         inp = pt.ones(4)
         _ = self.block(inp)
         mock.assert_called_once()
         actual = mock.call_args[0][0]
-        expected = pt.ones(4) * 4
+        expected = pt.ones(16) * 4
         pt.testing.assert_close(actual, expected)
 
-    def test_drop_called(self):
-        with patch.object(
-                self.block.drop,
-                'forward',
-                return_value=pt.ones(4)
-        ) as mock:
-            inp = pt.ones(4)
-            _ = self.block(inp)
-            mock.assert_called_once()
-            actual = mock.call_args[0][0]
-            expected = pt.ones(4) * 4
-            pt.testing.assert_close(actual, expected)
-
-    def test_widen_called(self):
-        with patch.object(
-                self.block.widen,
-                'forward',
-                return_value=pt.ones(8)
-        ) as mock:
-            inp = pt.ones(4)
-            _ = self.block(inp)
-            actual = mock.call_args[0][0]
-            expected = pt.ones(4) * 4
-            pt.testing.assert_close(actual, expected)
+    def test_shrink_called(self):
+        mock = Mock(return_value=pt.ones(8))
+        self.block.shrink.forward = mock
+        inp = pt.ones(4)
+        _ = self.block(inp)
+        mock.assert_called_once()
+        actual = mock.call_args[0][0]
+        expected = pt.ones(16) * 4
+        pt.testing.assert_close(actual, expected)
 
     def test_gate_called(self):
         mock = Mock(return_value=pt.ones(4))
@@ -221,8 +219,18 @@ class TestUsage(unittest.TestCase):
         _ = self.block(inp)
         mock.assert_called_once()
         actual = mock.call_args[0][0]
-        expected = pt.ones(4) * 16
+        expected = pt.ones(4) * 64
         pt.testing.assert_close(actual, expected)
+
+    def test_drop_called(self):
+        inp = pt.ones(4)
+        with patch.object(self.block.drop, 'forward') as mock:
+            mock.return_value=pt.ones(16)
+            _ = self.block(inp)
+            mock.assert_called_once()
+            actual = mock.call_args[0][0]
+            expected = pt.ones(16) * 4
+            pt.testing.assert_close(actual, expected)
 
 
 if __name__ == '__main__':