+++ tests/pt/mix/weighted/test_gated.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch
 import torch as pt
-from torch.nn import Linear, Softmax, GELU, Sigmoid
+from torch.nn import Linear, Softmax, GELU, Sigmoid, PReLU
 from swak.pt.misc import identity
 from swak.pt.mix.weighted import GatedSumMixer
 
@@ -69,9 +69,23 @@ class TestDefaultAttributes(unittest.TestCase):
         self.assertTrue(callable(self.mix.reset_parameters))
 
     @patch('torch.nn.Linear.reset_parameters')
+    def test_reset_parameters_called_on_instantiation(self, linear):
+        gate = PReLU()
+        with patch('torch.nn.PReLU.reset_parameters') as mock:
+            _ = GatedSumMixer(4, 3, gate)
+            self.assertEqual(1, mock.call_count)
+            self.assertEqual(1, linear.call_count)
+
+    @patch('torch.nn.Linear.reset_parameters')
     def test_reset_parameters_called(self, mock):
         self.mix.reset_parameters()
         mock.assert_called_once_with()
+
+    def test_reset_parameters_called_on_gate(self):
+        mix = GatedSumMixer(4, 3, PReLU())
+        with patch('torch.nn.PReLU.reset_parameters') as gate:
+            mix.reset_parameters()
+            self.assertEqual(1, gate.call_count)
 
     def test_has_new(self):
         self.assertTrue(hasattr(self.mix, 'new'))