+++ tests/pt/mix/weighted/test_gated_residual.py
@@ -2,6 +2,7 @@ import unittest
 from unittest.mock import patch, Mock
 import torch as pt
 from torch.nn import Linear, Softmax, GELU, Sigmoid, ELU, Dropout, AlphaDropout
+from torch.nn import PReLU
 from swak.pt.misc import identity
 from swak.pt.mix.weighted import GatedResidualSumMixer
 
@@ -91,9 +92,24 @@ class TestDefaultAttributes(unittest.TestCase):
         self.assertTrue(callable(self.mix.reset_parameters))
 
     @patch('torch.nn.Linear.reset_parameters')
+    def test_reset_parameters_called_on_instantiation(self, linear):
+        activate = PReLU()
+        gate = PReLU()
+        with patch('torch.nn.PReLU.reset_parameters') as mock:
+            _ = GatedResidualSumMixer(4, 3, activate, gate)
+            self.assertEqual(2, mock.call_count)
+            self.assertEqual(2, linear.call_count)
+
+    @patch('torch.nn.Linear.reset_parameters')
     def test_reset_parameters_called(self, mock):
         self.mix.reset_parameters()
         self.assertEqual(2, mock.call_count)
+
+    def test_reset_parameters_called_on_activations(self):
+        mix = GatedResidualSumMixer(4, 3, PReLU(), PReLU())
+        with patch('torch.nn.PReLU.reset_parameters') as activation:
+            mix.reset_parameters()
+            self.assertEqual(2, activation.call_count)
 
     def test_has_new(self):
         self.assertTrue(hasattr(self.mix, 'new'))